Create Common API module
