+++ digiroad2-api-common/src/main/scala/fi/liikennevirasto/digiroad2/authentication/UnauthenticatedException.scala
@@ -1,0 +1,3 @@
+package fi.liikennevirasto.digiroad2.authentication
+
+case class UnauthenticatedException() extends RuntimeException
