+++ digiroad2-api-common/src/main/scala/fi/liikennevirasto/digiroad2/authentication/UserNotFoundException.scala
@@ -1,0 +1,3 @@
+package fi.liikennevirasto.digiroad2.authentication
+
+case class UserNotFoundException(username: String) extends RuntimeException()
