--- src/main/scala/fi/liikennevirasto/digiroad2/authentication/UnauthenticatedException.scala
@@ -1,3 +1,0 @@
-package fi.liikennevirasto.digiroad2.authentication
-
-case class UnauthenticatedException() extends RuntimeException
