--- src/main/scala/fi/liikennevirasto/digiroad2/authentication/UserNotFoundException.scala
@@ -1,3 +1,0 @@
-package fi.liikennevirasto.digiroad2.authentication
-
-case class UserNotFoundException(username: String) extends RuntimeException()
